Fixed many resource file names, removed raw-char matches input-char requirement for Linux operation.
WARNING: This version is unstable. High possibility of encountering a complete processor lockup (likely an infinite loop).

Known bugs on Linux: Some fields are being drawn off-target (such as text) and they need to be fixed in order to allow compatibility. Different style files on different platforms would not be a good idea. The combined-editors are not working. When a text-replacement is made via "~", the cursor is placed before it and attempting to remove or select it causes a lockup. Symbol editor seems to be working fine. The symbol selection dialog causes a crash when used.

git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@194 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/gui/control/graph.cpp
@@ -161,12 +161,12 @@ void BarGraph::draw(RotatedDC& dc, int current) const {
 	}
 	// How many labels and lines to draw?
 	dc.SetFont(*wxNORMAL_FONT);
-	UInt label_step = max(1.0, dc.GetCharHeight() / step_height);
+	UInt label_step = UInt(max(1.0, dc.GetCharHeight() / step_height));
 	// Draw backlines (horizontal) and value labels
 	dc.SetPen(lerp(bg, fg, 0.5));
 	for (UInt i = 0 ; i <= axis.max ; ++i) {
 		if (i % label_step == 0) {
-			int y = rect.bottom() - i * step_height;
+			int y = int(rect.bottom() - i * step_height);
 			dc.DrawLine(RealPoint(rect.left() - 2, y), RealPoint(rect.right(), y));
 			// draw label, aligned middle right
 			String label; label << i;
@@ -202,7 +202,7 @@ int BarGraph::findItem(const RealPoint& pos, const RealRect& rect1) const {
 	double width_space = rect.width / count; // including spacing
 	double space       = width_space / 5;
 	// Find column in which this point could be located
-	int    col    = (pos.x - rect.x) / width_space;
+	int    col    = int((pos.x - rect.x) / width_space);
 	double in_col = (pos.x - rect.x) - col * width_space;
 	if (in_col < space / 2               || // left
 	    in_col > width_space - space / 2 || // right
